Add files via upload
this is a version of the test-code branch that I am starting on. its a basic login screen. i have implemented a method for when a user enters a password and its correct on the press of the key enter. password for your testing is "abc123". this is just until we get database functionality up and running, and the method likely needs to be moved to a different location in the overall program.